refactor: resolve config-telemetry circular reference
We refactor the interaction between telemetry and configuration to
resolve the circular import issue between these two components.
