+++ ddtrace/internal/appsec/product.py
@@ -1,4 +1,4 @@
-from ddtrace.settings.asm import config as asm_config
+from ddtrace.settings.asm import config
 
 
 requires = ["remote-configuration"]
@@ -9,14 +9,14 @@ def post_preload():
 
 
 def start():
-    if asm_config._asm_rc_enabled:
+    if config._asm_rc_enabled:
         from ddtrace.appsec._remoteconfiguration import enable_appsec_rc
 
         enable_appsec_rc()
 
 
 def restart(join=False):
-    if asm_config._asm_rc_enabled:
+    if config._asm_rc_enabled:
         from ddtrace.appsec._remoteconfiguration import _forksafe_appsec_rc
 
         _forksafe_appsec_rc()
