+++ ddtrace/internal/telemetry/writer.py
@@ -15,18 +15,16 @@ from typing import Union  # noqa:F401
 import urllib.parse as parse
 
 from ddtrace.internal.logger import get_logger
+from ddtrace.settings._telemetry import config
 
 from ...internal import atexit
 from ...internal import forksafe
-from ...settings._config import _get_config
-from ...settings._inferred_base_service import detect_service
 from ..agent import get_connection
 from ..agent import get_trace_url
 from ..encoding import JSONEncoderV2
 from ..periodic import PeriodicService
 from ..runtime import get_runtime_id
 from ..service import ServiceStatus
-from ..utils.formats import asbool
 from ..utils.time import StopWatch
 from ..utils.version import version as tracer_version
 from . import modules
@@ -50,34 +48,7 @@ from .metrics_namespaces import MetricNamespace
 from .metrics_namespaces import NamespaceMetricType  # noqa:F401
 
 
-_inferred_service = detect_service(sys.argv)
-
-
 log = get_logger(__name__)
-
-
-class _TelemetryConfig:
-    API_KEY: str = _get_config("DD_API_KEY", None, report_telemetry=False)
-    SITE: str = _get_config("DD_SITE", "datadoghq.com", report_telemetry=False)
-    ENV: str = _get_config("DD_ENV", "", report_telemetry=False)
-    SERVICE: str = _get_config("DD_SERVICE", _inferred_service or "unnamed-python-service", report_telemetry=False)
-    VERSION: str = _get_config("DD_VERSION", "", report_telemetry=False)
-    AGENTLESS_MODE: bool = _get_config("DD_CIVISIBILITY_AGENTLESS_ENABLED", False, asbool, report_telemetry=False)
-    DEBUG: bool = _get_config("DD_TRACE_DEBUG", False, asbool, report_telemetry=False)
-    HEARTBEAT_INTERVAL: float = _get_config("DD_TELEMETRY_HEARTBEAT_INTERVAL", 60, float, report_telemetry=False)
-    TELEMETRY_ENABLED: bool = _get_config("DD_INSTRUMENTATION_TELEMETRY_ENABLED", True, asbool, report_telemetry=False)
-    DEPENDENCY_COLLECTION: bool = _get_config(
-        "DD_TELEMETRY_DEPENDENCY_COLLECTION_ENABLED", True, asbool, report_telemetry=False
-    )
-    INSTALL_ID: Optional[str] = _get_config("DD_INSTRUMENTATION_INSTALL_ID", None, report_telemetry=False)
-    INSTALL_TYPE: Optional[str] = _get_config("DD_INSTRUMENTATION_INSTALL_TYPE", None, report_telemetry=False)
-    INSTALL_TIME: Optional[str] = _get_config("DD_INSTRUMENTATION_INSTALL_TIME", None, report_telemetry=False)
-    FORCE_START: bool = _get_config(
-        "_DD_INSTRUMENTATION_TELEMETRY_TESTS_FORCE_APP_STARTED", False, asbool, report_telemetry=False
-    )
-    LOG_COLLECTION_ENABLED: bool = TELEMETRY_ENABLED and _get_config(
-        "DD_TELEMETRY_LOG_COLLECTION_ENABLED", True, asbool, report_telemetry=False
-    )
 
 
 class LogData(dict):
@@ -99,7 +70,7 @@ class _TelemetryClient:
 
     def __init__(self, agentless):
         # type: (bool) -> None
-        self._telemetry_url = self.get_host(_TelemetryConfig.SITE, agentless)
+        self._telemetry_url = self.get_host(config.SITE, agentless)
         self._endpoint = self.get_endpoint(agentless)
         self._encoder = JSONEncoderV2()
         self._agentless = agentless
@@ -110,8 +81,8 @@ class _TelemetryClient:
             "DD-Client-Library-Version": tracer_version,
         }
 
-        if agentless and _TelemetryConfig.API_KEY:
-            self._headers["dd-api-key"] = _TelemetryConfig.API_KEY
+        if agentless and config.API_KEY:
+            self._headers["dd-api-key"] = config.API_KEY
 
     @property
     def url(self):
@@ -181,11 +152,11 @@ class TelemetryWriter(PeriodicService):
 
     def __init__(self, is_periodic=True, agentless=None):
         # type: (bool, Optional[bool]) -> None
-        super(TelemetryWriter, self).__init__(interval=min(_TelemetryConfig.HEARTBEAT_INTERVAL, 10))
+        super(TelemetryWriter, self).__init__(interval=min(config.HEARTBEAT_INTERVAL, 10))
         # Decouples the aggregation and sending of the telemetry events
         # TelemetryWriter events will only be sent when _periodic_count == _periodic_threshold.
         # By default this will occur at 10 second intervals.
-        self._periodic_threshold = int(_TelemetryConfig.HEARTBEAT_INTERVAL // self.interval) - 1
+        self._periodic_threshold = int(config.HEARTBEAT_INTERVAL // self.interval) - 1
         self._periodic_count = 0
         self._is_periodic = is_periodic
         self._integrations_queue = dict()  # type: Dict[str, Dict]
@@ -209,14 +180,14 @@ class TelemetryWriter(PeriodicService):
         self.started = False
 
         # Debug flag that enables payload debug mode.
-        self._debug = _TelemetryConfig.DEBUG
+        self._debug = config.DEBUG
 
-        self._enabled = _TelemetryConfig.TELEMETRY_ENABLED
+        self._enabled = config.TELEMETRY_ENABLED
 
         if agentless is None:
-            agentless = _TelemetryConfig.AGENTLESS_MODE or _TelemetryConfig.API_KEY not in (None, "")
+            agentless = config.AGENTLESS_MODE or config.API_KEY not in (None, "")
 
-        if agentless and not _TelemetryConfig.API_KEY:
+        if agentless and not config.API_KEY:
             log.debug("Disabling telemetry: no Datadog API key found in agentless mode")
             self._enabled = False
         self._client = _TelemetryClient(agentless)
@@ -234,9 +205,9 @@ class TelemetryWriter(PeriodicService):
             # This will occur when the agent writer starts.
             self.enable()
             # Force app started for unit tests
-            if _TelemetryConfig.FORCE_START:
+            if config.FORCE_START:
                 self._app_started()
-            if _TelemetryConfig.LOG_COLLECTION_ENABLED:
+            if config.LOG_COLLECTION_ENABLED:
                 get_logger("ddtrace").addHandler(DDTelemetryLogHandler(self))
 
     def enable(self):
@@ -297,9 +268,7 @@ class TelemetryWriter(PeriodicService):
                 "api_version": "v2",
                 "seq_id": next(self._sequence),
                 "debug": self._debug,
-                "application": get_application(
-                    _TelemetryConfig.SERVICE, _TelemetryConfig.VERSION, _TelemetryConfig.ENV
-                ),
+                "application": get_application(config.SERVICE, config.VERSION, config.ENV),
                 "host": get_host_info(),
                 "payload": payload,
                 "request_type": payload_type,
@@ -365,11 +334,11 @@ class TelemetryWriter(PeriodicService):
             "products": products,
         }  # type: Dict[str, Union[Dict[str, Any], List[Any]]]
         # Add time to value telemetry metrics for single step instrumentation
-        if _TelemetryConfig.INSTALL_ID or _TelemetryConfig.INSTALL_TYPE or _TelemetryConfig.INSTALL_TIME:
+        if config.INSTALL_ID or config.INSTALL_TYPE or config.INSTALL_TIME:
             payload["install_signature"] = {
-                "install_id": _TelemetryConfig.INSTALL_ID,
-                "install_type": _TelemetryConfig.INSTALL_TYPE,
-                "install_time": _TelemetryConfig.INSTALL_TIME,
+                "install_id": config.INSTALL_ID,
+                "install_type": config.INSTALL_TYPE,
+                "install_time": config.INSTALL_TIME,
             }
 
         # Reset the error after it has been reported.
@@ -378,10 +347,7 @@ class TelemetryWriter(PeriodicService):
 
     def _app_heartbeat_event(self):
         # type: () -> None
-        if (
-            _TelemetryConfig.DEPENDENCY_COLLECTION
-            and time.monotonic() - self._extended_time > self._extended_heartbeat_interval
-        ):
+        if config.DEPENDENCY_COLLECTION and time.monotonic() - self._extended_time > self._extended_heartbeat_interval:
             self._extended_time += self._extended_heartbeat_interval
             self._app_dependencies_loaded_event()
             payload = {
@@ -437,7 +403,7 @@ class TelemetryWriter(PeriodicService):
     def _app_dependencies_loaded_event(self):
         """Adds events to report imports done since the last periodic run"""
 
-        if not _TelemetryConfig.DEPENDENCY_COLLECTION or not self._enabled:
+        if not config.DEPENDENCY_COLLECTION or not self._enabled:
             return
         with self._service_lock:
             newly_imported_deps = modules.get_newly_imported_modules(self._modules_already_imported)
