+++ ddtrace/settings/__init__.py
@@ -1,12 +1,10 @@
 from .._hooks import Hooks
-from ._config import Config
 from .exceptions import ConfigException
 from .http import HttpConfig
 from .integration import IntegrationConfig
 
 
 __all__ = [
-    "Config",
     "ConfigException",
     "HttpConfig",
     "Hooks",
