+++ ddtrace/settings/dynamic_instrumentation.py
@@ -7,7 +7,6 @@ from ddtrace.internal.agent import get_trace_url
 from ddtrace.internal.constants import DEFAULT_SERVICE_NAME
 from ddtrace.internal.utils.config import get_application_name
 from ddtrace.settings._core import DDConfig
-from ddtrace.settings._telemetry import report_telemetry as _report_telemetry
 from ddtrace.version import get_version
 
 
@@ -131,4 +130,3 @@ class DynamicInstrumentationConfig(DDConfig):
 
 
 config = DynamicInstrumentationConfig()
-_report_telemetry(config)
