+++ ddtrace/settings/peer_service.py
@@ -1,8 +1,8 @@
 from ddtrace.ext import SpanKind
 from ddtrace.internal.schema import SCHEMA_VERSION
+from ddtrace.internal.telemetry import get_config as _get_config
 from ddtrace.internal.utils.formats import asbool
 from ddtrace.internal.utils.formats import parse_tags_str
-from ddtrace.settings._core import get_config as _get_config
 
 
 class PeerServiceConfig(object):
