+++ tests/ci_visibility/util.py
@@ -14,6 +14,7 @@ from ddtrace.internal.ci_visibility.git_client import CIVisibilityGitClient
 from ddtrace.internal.ci_visibility.recorder import CIVisibility
 from ddtrace.internal.ci_visibility.recorder import CIVisibilityTracer
 from ddtrace.internal.test_visibility._internal_item_ids import InternalTestId
+from ddtrace.settings._config import Config
 from tests.utils import DummyCIVisibilityWriter
 from tests.utils import override_env
 
@@ -29,7 +30,8 @@ def _patch_dummy_writer():
 def _get_default_civisibility_ddconfig(itr_skipping_level: ITR_SKIPPING_LEVEL = ITR_SKIPPING_LEVEL.TEST):
     if not isinstance(itr_skipping_level, ITR_SKIPPING_LEVEL):
         raise ValueError(f"Invalid ITR_SKIPPING_LEVEL: {itr_skipping_level}")
-    new_ddconfig = ddtrace.settings.Config()
+
+    new_ddconfig = Config()
     new_ddconfig._add(
         "test_visibility",
         {
