+++ tests/contrib/dbapi/test_dbapi_appsec.py
@@ -3,7 +3,7 @@ import pytest
 
 from ddtrace.appsec._iast import oce
 from ddtrace.contrib.dbapi import TracedCursor
-from ddtrace.settings import Config
+from ddtrace.settings._config import Config
 from ddtrace.settings.asm import config as asm_config
 from ddtrace.settings.integration import IntegrationConfig
 from ddtrace.trace import Pin
